refactor(DrawLottoNum) : 로또 클래스를 그대로 사용하기로 결정
기존의 하드코딩되어 있던 부분 상수화
